requirements/base.txt
# Base
psycopg2-binary==2.9.3

# Django
Django==4.0.4

# Environment
django-environ==0.8.1

# Django REST Frmework
djangorestframework==3.13.1
django-filter==2.4.0

# Utilis
geopy==2.2.0
requests==2.27.1

# Graphql
django-restql==0.5.3

# Celery
redis>=3.2.0
django-redis==4.10.0
celery==4.3.0
tornado>=5.0.0,<6.0.0
flower==0.9.2
